完善文件上传下载

+++ src/main/java/com/mq/jobManagement/back_end/controller/StudentController.java
@@ -12,6 +12,6 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 @RequestMapping("/student")
-public class StudentController extends CommonController<StudentService, Student,Integer> {
+public class StudentController extends CommonController<StudentService, Student,Long> {
 
 }
