+++ src/main/java/com/mq/jobManagement/back_end/pojo/Student.java
@@ -16,6 +16,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @TableName("student")
-public class Student extends CommonEntity<Integer> {
+public class Student extends CommonEntity<Long> {
     private String name;
 }
